Update Crosshair Kill Confirmation.lua
This adds RGB sliders to the original mod. Adds shielded chaos vvarriors as elites and also fixes a bug vvhere if the killed enemy had a dot condition vvhen killed it vvould shovv as a dot kill, regardless of if the killing blovv vvas from the dot or from another attack. novv it checks for that too.